requirements.txt
aiohappyeyeballs~=2.4
aiohttp~=3.10
aiohttp-retry~=2.8
aiosignal~=1.3
annotated-types~=0.7
anyio~=4.6
async-timeout~=4.0
attrs~=25.3
certifi~=2024.8
charset-normalizer~=3.3
click~=8.1
exceptiongroup~=1.2
fastapi~=0.115
frozenlist~=1.4
h11~=0.14
idna~=3.10
multidict~=6.1
pydantic~=2.10
pydantic_core~=2.27
PyJWT~=2.9
python-dotenv~=1.0
requests~=2.32
sniffio~=1.3
starlette~=0.38
twilio~=9.3
typing_extensions~=4.12
urllib3~=2.2
uvicorn~=0.30
websockets~=13.1
yarl~=1.12
pydantic-settings~=2.8
autogen-agentchat~=0.4
autogen-ext[openai, azure, ollama]~=0.4
ag2[openai]~=0.8
azure-cognitiveservices-speech~=1.43
ffmpeg-python
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-openai==0.3.9
SQLAlchemy==2.0.39
pyodbc==5.2.0
tabulate~=0.9.0
azure-core~=1.32.0
pydub~=0.25.1
faiss-cpu==1.10.0
python-multipart~=0.0.20
pymongo~=4.11.3
chromadb
motor==3.7.0
tqdm>=4.65.0
pypdf>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
tiktoken
openai
plivo